YLP-2446 Patient does not have to retrieve a list of patients
