+++ packages/yourloops/components/patient-list/current-patient-list/private-team-or-caregiver-patient-list/private-team-or-caregiver-patient-list.tsx
@@ -34,7 +34,7 @@ import { EmptyPatientList } from '../../empty-patient-list/empty-patient-list'
 import { PatientListCustomFooter } from '../../patient-list-custom-footer'
 import RemoveDirectShareDialog from '../../../dialogs/remove-direct-share-dialog'
 import { usePatientListContext } from '../../../../lib/providers/patient-list.provider'
-import { usePatientContext } from '../../../../lib/patient/patient.provider'
+import { usePatientsContext } from '../../../../lib/patient/patients.provider'
 import { useWindowDimensions } from '../../../../lib/custom-hooks/use-window-dimensions.hook'
 import { PatientListColumns } from '../../models/enums/patient-list.enum'
 import RemovePatientDialog from '../../../patient/remove-patient-dialog/remove-patient-dialog'
@@ -54,7 +54,7 @@ export const PrivateTeamOrCaregiverPatientList: FunctionComponent<PrivateTeamOrC
     onRowClick
   } = usePrivateTeamOrCaregiverPatientListHook({ patients })
   const { displayedColumns } = usePatientListContext()
-  const { refreshInProgress } = usePatientContext()
+  const { refreshInProgress } = usePatientsContext()
   const { width } = useWindowDimensions()
   const gridApiRef = useGridApiRef()
