+++ packages/yourloops/components/patient/reinvite-patient-dialog/reinvite-patient-dialog.hook.tsx
@@ -28,7 +28,7 @@
 import { useTranslation } from 'react-i18next'
 import { useState } from 'react'
 import { useAlert } from '../../utils/snackbar'
-import { usePatientContext } from '../../../lib/patient/patient.provider'
+import { usePatientsContext } from '../../../lib/patient/patients.provider'
 import { type Patient } from '../../../lib/patient/models/patient.model'
 import { useSelectedTeamContext } from '../../../lib/selected-team/selected-team.provider'
 
@@ -48,7 +48,7 @@ export const useReinvitePatientDialog = ({
 }: ReinvitePatientDialogHookProps): ReinvitePatientDialogHookReturn => {
   const { t } = useTranslation('yourloops')
   const alert = useAlert()
-  const { removePatient, invitePatient } = usePatientContext()
+  const { removePatient, invitePatient } = usePatientsContext()
   const { selectedTeam } = useSelectedTeamContext()
 
   const [processing, setProcessing] = useState<boolean>(false)
