+++ packages/yourloops/test/unit/components/team/leave-team-button.test.tsx
@@ -29,7 +29,7 @@ import React from 'react'
 import { act, fireEvent, render, screen, waitFor, within } from '@testing-library/react'
 
 import * as teamHookMock from '../../../../lib/team'
-import * as patientHookMock from '../../../../lib/patient/patient.provider'
+import * as patientsHookMock from '../../../../lib/patient/patients.provider'
 import * as authHookMock from '../../../../lib/auth'
 import * as alertHookMock from '../../../../components/utils/snackbar'
 import { buildTeam, buildTeamMember } from '../../common/utils'
@@ -58,7 +58,7 @@ describe('TeamMembers', () => {
     (teamHookMock.useTeam as jest.Mock).mockImplementation(() => {
       return { leaveTeam: leaveTeamMock, getTeam: jest.fn().mockReturnValue(team) }
     });
-    (patientHookMock.usePatientContext as jest.Mock).mockImplementation(() => {
+    (patientsHookMock.usePatientsContext as jest.Mock).mockImplementation(() => {
       return { leaveTeam: patientLeaveTeamMock }
     });
     (authHookMock.useAuth as jest.Mock).mockImplementation(() => ({
